+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change short CLI argument for development from `-d` to `-D`.
+
 ## [0.16.0] - 2024-03-19
 
 ### Changed